add test runs api
