+++ py_jama_client/client.py
@@ -396,50 +396,3 @@ class JamaClient(BaseClient):
             params=params,
             allowed_results_per_page=allowed_results_per_page,
         )
-
-    def get_testruns(
-        self,
-        test_cycle_id: int,
-        *args,
-        params: Optional[dict] = None,
-        allowed_results_per_page=DEFAULT_ALLOWED_RESULTS_PER_PAGE,
-        **kwargs,
-    ):
-        """
-        This method will return all test runs associated with the specified test cycle.  Test runs will be returned
-        as a list of json objects.
-        Args:
-            test_cycle_id: (int) The id of the test cycle
-        """
-        resource_path = f"testcycles/{test_cycle_id}/testruns"
-        return self.get_all(
-            resource_path,
-            params,
-            allowed_results_per_page=allowed_results_per_page,
-            **kwargs,
-        )
-
-    def put_test_run(
-        self,
-        test_run_id: int,
-        data: dict = None,
-        *args,
-        params: Optional[dict] = None,
-        **kwargs,
-    ):
-        """This method will post a test run to Jama through the API"""
-        resource_path = f"testruns/{test_run_id}"
-        headers = {"content-type": "application/json"}
-        try:
-            response = self._core.put(
-                resource_path,
-                params,
-                data=data,
-                headers=headers,
-                **kwargs,
-            )
-        except CoreException as err:
-            py_jama_rest_client_logger.error(err)
-            raise APIException(str(err))
-        self.handle_response_status(response)
-        return response.status_code
